Use different scale factors in portrait or landscape

+++ draggablepanel/src/main/java/com/github/pedrovgs/DraggablePanel.java
@@ -76,7 +76,7 @@ public class DraggablePanel extends FrameLayout {
         this.fragmentManager = fragmentManager;
     }
 
-    public void setScaleFactor(int scaleFactor) {
+    public void setScaleFactor(float scaleFactor) {
         this.scaleFactor = scaleFactor;
     }
